@@ -24,4 +24,4 @@ const nextConfig = {
   },
 }
 
-module.exports = nextConfig
+export default nextConfig
